feat: add self-update command
