+++ pnpm-workspace.yaml
@@ -24,6 +24,7 @@ packages:
   - pkg-manifest/*
   - patching/*
   - pnpm
+  - tools/*
   - worker
   - pnpm/artifacts/*
   - releasing/*
