+++ tools/path/README.md
@@ -1,0 +1,15 @@
+# @pnpm/tools.path
+
+> Path to tools
+
+[![npm version](https://img.shields.io/npm/v/@pnpm/tools.path.svg)](https://www.npmjs.com/package/@pnpm/tools.path)
+
+## Installation
+
+```
+pnpm add @pnpm/tools.path
+```
+
+## License
+
+MIT
