+++ tools/path/src/index.ts
@@ -1,0 +1,13 @@
+import path from 'path'
+
+export function getToolDirPath (
+  opts: {
+    pnpmHomeDir: string
+    tool: {
+      name: string
+      version: string
+    }
+  }
+) {
+  return path.join(opts.pnpmHomeDir, '.tools', opts.tool.name.replaceAll('/', '+'), opts.tool.version)
+}
